+++ brawlstats/__init__.py
@@ -1,6 +1,5 @@
-from .core import Client
+from .brawlapi import Client as BrawlAPI
 from .errors import *
-from .models import *
 
 
 ############
@@ -8,7 +7,7 @@ from .models import *
 ############
 
 
-__version__ = 'v2.3.13'
+__version__ = 'v3.0.0'
 __title__ = 'brawlstats'
 __license__ = 'MIT'
 __author__ = 'SharpBit'